Keep frame references valid until end of layout loop
Previously the frame stack stored actual values that were popped once the element
was closed and it could not be retrieved again. Now the frame stack contains references
to elements and the list of actual elements is kept separately and can be referenced even
after closing the element.

Pros: Elements can be accessed after they're closed; smaller memory footprint.
Cons: Previously only a small list of elements needed to be kept (max number of nestable elements),
now the pool must be larger. Also getting the new element using CIG macro is a bit cumbersome.

This is a step towards building some sort of an API to pin edges to existing
elements and defining position and size of a new element that way.