[Core] Improves initial clip rect handling
Does not populate initial clip rects stack when starting a new layout. Previously there was ambiguity because
root element wasn't actually clipped. Now it's clear the root is not clipped, and calling cig_enable_clipping enables.
Boolean option in the callback method was renamed to 'reset' to indicate clipping was disabled for this buffer.